refactored dialogs entirely for owner
